+++ shm/src/Cargo.cpp
@@ -1,11 +1,11 @@
 #include "../inc/Cargo.hpp"
 
-Cargo& Cargo::operator+=(const uint16_t value){
+Cargo& Cargo::operator+=(const size_t value){
     amount_ += value;
     return *this;
 }
 
-Cargo& Cargo::operator-=(const uint16_t value){
+Cargo& Cargo::operator-=(const size_t value){
     amount_ -= value;
     return *this;
 }
