+++ shm/src/Ship.cpp
@@ -1,6 +1,6 @@
 #include "../inc/Ship.hpp"
 
-Ship::Ship(int8_t id, std::string name, uint8_t speed, uint8_t maxCrew, uint8_t capacity)
+Ship::Ship(size_t id, std::string name, size_t speed, size_t maxCrew, size_t capacity)
     : id_(id)
     , name_(name)
     , speed_(speed)
@@ -8,16 +8,16 @@ Ship::Ship(int8_t id, std::string name, uint8_t speed, uint8_t maxCrew, uint8_t 
     , capacity_(capacity)
 {}
 
-Ship::Ship(int8_t id, uint8_t speed, uint8_t maxCrew) 
+Ship::Ship(size_t id, size_t speed, size_t maxCrew) 
     : Ship(id, "Ship", speed, maxCrew, 100)
 {}
 
-Ship& Ship::operator+=(const uint16_t value){
+Ship& Ship::operator+=(const size_t value){
     crew_ += value;
     return *this;
 }
 
-Ship& Ship::operator-=(const uint16_t value){
+Ship& Ship::operator-=(const size_t value){
     crew_ -= value;
     return *this;
 }
